Fixed freezing issue in multi-tasking menu.
When the back button is pressed, the app freezes and crashes. This
happened because oslDeleteImage removed the image before we entered into
the multi-tasking menu. And therefore when the back button is hit it
would return to the previous menu, and the image was deleted. Since the
image was deleted  upon returning, the app crashes and freezes.